+++ opentelemetry-sdk/src/opentelemetry/sdk/metrics/_internal/aggregation.py
@@ -471,15 +471,15 @@ class _ExplicitBucketHistogramAggregation(_Aggregation[HistogramPoint]):
                     is AggregationTemporality.DELTA
                 ):
 
-                    if current_value is None:
-                        return None
-
                     previous_collection_start_nano = (
                         self._previous_collection_start_nano
                     )
                     self._previous_collection_start_nano = (
                         collection_start_nano
                     )
+
+                    if current_value is None:
+                        return None
 
                     return HistogramDataPoint(
                         attributes=self._attributes,