+++ opentelemetry-sdk/tests/metrics/test_point.py
@@ -178,7 +178,7 @@ class TestToJson(TestCase):
             metrics=[cls.metric_0, cls.metric_1, cls.metric_2],
             schema_url="schema_url_0",
         )
-        cls.scope_metrics_0_str = f'{{"scope": {{"name": "name_0", "version": "version_0", "schema_url": "schema_url_0"}}, "metrics": [{cls.metric_0_str}, {cls.metric_1_str}, {cls.metric_2_str}], "schema_url": "schema_url_0"}}'
+        cls.scope_metrics_0_str = f'{{"scope": {{"name": "name_0", "version": "version_0", "schema_url": "schema_url_0", "attributes": null}}, "metrics": [{cls.metric_0_str}, {cls.metric_1_str}, {cls.metric_2_str}], "schema_url": "schema_url_0"}}'
 
         cls.scope_metrics_1 = ScopeMetrics(
             scope=InstrumentationScope(
@@ -189,7 +189,7 @@ class TestToJson(TestCase):
             metrics=[cls.metric_0, cls.metric_1, cls.metric_2],
             schema_url="schema_url_1",
         )
-        cls.scope_metrics_1_str = f'{{"scope": {{"name": "name_1", "version": "version_1", "schema_url": "schema_url_1"}}, "metrics": [{cls.metric_0_str}, {cls.metric_1_str}, {cls.metric_2_str}], "schema_url": "schema_url_1"}}'
+        cls.scope_metrics_1_str = f'{{"scope": {{"name": "name_1", "version": "version_1", "schema_url": "schema_url_1", "attributes": null}}, "metrics": [{cls.metric_0_str}, {cls.metric_1_str}, {cls.metric_2_str}], "schema_url": "schema_url_1"}}'
 
         cls.resource_metrics_0 = ResourceMetrics(
             resource=Resource(